rewrite modint
